+++ include/readout/models/SkipListLatencyBufferModel.hpp
@@ -12,7 +12,7 @@
 #ifndef READOUT_INCLUDE_READOUT_MODELS_SKIPLISTLATENCYBUFFERMODEL_HPP_
 #define READOUT_INCLUDE_READOUT_MODELS_SKIPLISTLATENCYBUFFERMODEL_HPP_
 
-#include "ReadoutIssues.hpp"
+#include "readout/ReadoutIssues.hpp"
 #include "readout/concepts/LatencyBufferConcept.hpp"
 
 #include "folly/ConcurrentSkipList.h"
