+++ include/readout/models/SourceEmulatorModel.hpp
@@ -17,7 +17,7 @@
 
 #include "readout/fakecardreader/Structs.hpp"
 
-#include "ReadoutIssues.hpp"
+#include "readout/ReadoutIssues.hpp"
 #include "readout/concepts/SourceEmulatorConcept.hpp"
 #include "readout/utils/FileSourceBuffer.hpp"
 #include "readout/utils/RateLimiter.hpp"
