+++ include/readout/utils/BufferedFileReader.hpp
@@ -9,7 +9,7 @@
 #ifndef READOUT_INCLUDE_READOUT_UTILS_BUFFEREDFILEREADER_HPP_
 #define READOUT_INCLUDE_READOUT_UTILS_BUFFEREDFILEREADER_HPP_
 
-#include "ReadoutIssues.hpp"
+#include "readout/ReadoutIssues.hpp"
 #include "readout/ReadoutLogging.hpp"
 
 #include "logging/Logging.hpp"
