+++ include/readout/utils/FileSourceBuffer.hpp
@@ -8,7 +8,7 @@
 #ifndef READOUT_INCLUDE_READOUT_UTILS_FILESOURCEBUFFER_HPP_
 #define READOUT_INCLUDE_READOUT_UTILS_FILESOURCEBUFFER_HPP_
 
-#include "ReadoutIssues.hpp"
+#include "readout/ReadoutIssues.hpp"
 #include "readout/ReadoutLogging.hpp"
 
 #include "logging/Logging.hpp"
