+++ plugins/FakeCardReader.cpp
@@ -11,7 +11,7 @@
 
 #include "CreateSourceEmulator.hpp"
 #include "FakeCardReader.hpp"
-#include "ReadoutIssues.hpp"
+#include "readout/ReadoutIssues.hpp"
 #include "readout/models/SourceEmulatorModel.hpp"
 
 #include "appfwk/app/Nljs.hpp"
