+++ src/daphne/DAPHNEFrameProcessor.hpp
@@ -8,7 +8,7 @@
 #ifndef READOUT_SRC_DAPHNE_DAPHNEFRAMEPROCESSOR_HPP_
 #define READOUT_SRC_DAPHNE_DAPHNEFRAMEPROCESSOR_HPP_
 
-#include "ReadoutIssues.hpp"
+#include "readout/ReadoutIssues.hpp"
 #include "readout/models/TaskRawDataProcessorModel.hpp"
 
 #include "dataformats/daphne/DAPHNEFrame.hpp"
